Splits CPU and CUDA fusion passes
