+++ torch/csrc/jit/fusers/cpu/cpu_fusion_pass.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "torch/csrc/jit/ir.h"
+
+namespace torch { namespace jit {
+
+// NB: Be sure to run DCE before fusion, because dead instructions
+// can prevent fusion opportunities from being exploited.
+TORCH_API void FuseCPUGraph(std::shared_ptr<Graph>& graph);
+
+} // namespace jit
+} // namespace torch
